Create README.md
The purpose of the UD Annotatrix is to provide interfaces to the current [visualiser tool](http://jonorthwash.github.io/visualise.html) that allows dependency graphs to be modified.  The point of this is to make manual editing of dependency corpora much quicker.

See [examples](http://universaldependencies.org/kk/dep/conj.html) of UD graphs using the same library as the visualiser currently does use and as the annotatrix will use.

Note that something similar exists in brat ([example](http://kazcorpus.kz/brat1/#/_qq01/_qq_005_109)), but that we're aiming for a simpler, cleaner, faster interface that has no server-side component.  Everything should be in javascript.