Fix #381: prevent corpus modifications from resetting the index
If two users were editing different sentences on the same treebank
in different windows connected to the same backend, then making a
change in one window would cause the `app.corpus.index` on the other
window to be reset (set to 0).  This meant that whenever one user
made a change, the other user would have to navigate all the way
back to their place.

This bug was happening because whenever we re-loaded the corpus via
the `/save` ajax path, we would broadcast an "index" of 0 to all the
other connected users, since the `/save` path is meant for entire-
corpus overwrites.  The app would then interpret this index as the
current index for the window, and then reset it.

This patch introduces a simple fix: save the current index before
we overwrite the corpus, and then just navigate to it after overwriting.

Also included are rebuilt bundles and an updated package-lock.json.