App: Never try to download a "null" format
If the sentence is not "parsed", then it's possible that we could
have `app.corpus.format` return `null`.  However, the `Sentence::to()`
method expects that its `format` argument will be the key of a mapping
it maintains to generator functions.  In the case of a `null` format,
we'll get an "undefined is not a function"-type error and die.

This patch makes it so an "unparsed" sentence will always just spit
out plain text.  (Ideally this should also warn the user or
something...)