client: Remove unused "test" code
This code hasn't been functional since at least 67eb0ff (Jan 2019),
and probably even 0262c13 (Aug 2018).  There are probably _some_ test
cases here that would be worth salvaging, but the vast majority of it
relies on deleted APIs...

If it turns out that we do want to come back and actually run some
tests, I think it'll be easier to start with a clean slate, using the
stuff removed in this patch as inspiration.